sync

+++ Khazad/DF Extractor/DfMap.h
@@ -157,8 +157,8 @@ private:
     uint32_t blocks_read;
     bool valid;
 
-    void convertToDfMapCoords(int x, int y, int &out_x, int &out_y, int &out_x2, int &out_y2);
-    void allocBlockArray(int x,int y, int z);
+    void convertToDfMapCoords(uint32_t x, uint32_t y, uint32_t &out_x, uint32_t &out_y, uint32_t &out_x2, uint32_t &out_y2);
+    void allocBlockArray(uint32_t x,uint32_t y, uint32_t z);
     void updateCellCount();
     uint32_t getBlocksCount() const { return blocks_read; }
     bool loadVersion1(FILE * Decompressed,DfMapHeader & h);
@@ -193,9 +193,9 @@ public:
     bool write(string FilePath);
     void clear();
 
-    Block* getBlock(int x, int y, int z);
-    Block* allocBlock(int x, int y, int z);
-    bool   deallocBlock(int x, int y, int z);
+    Block* getBlock(uint32_t x, uint32_t y, uint32_t z);
+    Block* allocBlock(uint32_t x, uint32_t y, uint32_t z);
+    bool   deallocBlock(uint32_t x, uint32_t y, uint32_t z);
 
     unsigned x_block_count, y_block_count, z_block_count; // block count
     unsigned x_cell_count, y_cell_count, z_cell_count;    // cell count
@@ -207,34 +207,34 @@ public:
     unsigned int getYBlocks()        { return y_block_count; }
     unsigned int getZBlocks()        { return z_block_count; }
 
-    bool isTileSky(int x, int y, int z, int blockX, int blockY);
-    short int getTileType(int x, int y, int z);
-    short int getTileType(int x, int y, int z, int blockX, int blockY);
+    bool isTileSky(uint32_t x, uint32_t y, uint32_t z, uint32_t blockX, uint32_t blockY);
+    uint16_t getTileType(uint32_t x, uint32_t y, uint32_t z);
+    uint16_t getTileType(uint32_t x, uint32_t y, uint32_t z, uint32_t blockX, uint32_t blockY);
 
-    int getDesignations(int x, int y, int z);
-    int getOccupancies(int x, int y, int z);
+    uint32_t getDesignations(uint32_t x, uint32_t y, uint32_t z);
+    uint32_t getOccupancies(uint32_t x, uint32_t y, uint32_t z);
 
     // get tile material
-    int16_t getMaterialIndex (int x, int y, int z);
-    string getMaterialString (int x, int y, int z);
+    int16_t getMaterialIndex (uint32_t x, uint32_t y, uint32_t z);
+    string getMaterialString (uint32_t x, uint32_t y, uint32_t z);
 
     // get coords of region used for materials
-    void getGeoRegion (int x, int y, int z, int& geoX, int& geoY);
+    void getGeoRegion (uint32_t x, uint32_t y, uint32_t z, int32_t& geoX, int32_t& geoY);
 
     // matgloss part of the designation
-    unsigned int getGeolayerIndex (int x, int y, int z);
+    uint32_t getGeolayerIndex (uint32_t x, uint32_t y, uint32_t z);
 
     // terrible
-    uint32_t getGeolayerAddress (int x, int y, int z);
-    uint32_t getGeoblockAddress (int x, int y, int z);
-    uint32_t getRegionAddress(int x, int y, int z);
+    uint32_t getGeolayerAddress (uint32_t x, uint32_t y, uint32_t z);
+    uint32_t getGeoblockAddress (uint32_t x, uint32_t y, uint32_t z);
+    uint32_t getRegionAddress(uint32_t x, uint32_t y, uint32_t z);
 
-    void getRegionCoords (int &x,int &y,int &z);
-    void setRegionCoords (int x,int y,int z);
+    void getRegionCoords (uint32_t &x,uint32_t &y,uint32_t &z);
+    void setRegionCoords (uint32_t x,uint32_t y,uint32_t z);
 
     // what kind of building is here?
-    uint16_t getBuilding (int x, int y, int z);
-    unsigned int getBiome (int x, int y, int z);
+    uint16_t getBuilding (uint32_t x, uint32_t y, uint32_t z);
+    unsigned int getBiome (uint32_t x, uint32_t y, uint32_t z);
 
     int picktexture(int);
 
@@ -244,15 +244,15 @@ public:
     bool isFloorTerrain(int);
     bool isWallTerrain(int);
 
-    bool isBlockInitialized(int x, int y, int z);
+    bool isBlockInitialized(uint32_t x, uint32_t y, uint32_t z);
 
-    bool isHidden (int x, int y, int z);
-    bool isSubterranean (int x, int y, int z);
-    bool isSkyView (int x, int y, int z);
-    bool isSunLit (int x, int y, int z);
-    bool isMagma (int x, int y, int z);
+    bool isHidden (uint32_t x, uint32_t y, uint32_t z);
+    bool isSubterranean (uint32_t x, uint32_t y, uint32_t z);
+    bool isSkyView (uint32_t x, uint32_t y, uint32_t z);
+    bool isSunLit (uint32_t x, uint32_t y, uint32_t z);
+    bool isMagma (uint32_t x, uint32_t y, uint32_t z);
 
-    int getLiquidLevel(int x, int y, int z);
+    uint8_t getLiquidLevel(uint32_t x, uint32_t y, uint32_t z);
 };
